component with errors
