+++ src/components/applicationAreaForm/index.js
@@ -1,0 +1,2 @@
+import { ApplicationAreaForm } from './applicationAreaForm';
+export default ApplicationAreaForm;
