+++ src/pages/addApplicationArea/addApplicationArea.tsx
@@ -1,5 +1,5 @@
 import { MainContainer } from "../../components/mainContainer/MainContainer";
-import { TextField } from "@mui/material";
+import { ApplicationAreaForm } from "../../components/applicationAreaForm/applicationAreaForm";
 import { useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import "./addApplicationArea.css";
@@ -61,6 +61,8 @@ export const AddApplicationArea = () => {
     const backendErrors: BackendError[] = [];
     verifyFormErrors(formErrors);
 
+    console.log('formerrores length', formErrors);
+    
     if (formErrors.length === 0) {
       const response: any = await createApplicationArea(formatterForm());
       console.log("respuesta back", response);
@@ -81,24 +83,6 @@ export const AddApplicationArea = () => {
     setFormErrors(formErrors);
   };
 
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const errors: FormError[] = [];
-    const fieldName = event.target?.name; // valor de la propiedad name del input
-    const validationMessage = event.target?.validationMessage; // mensaje de error de validaciones que no son de backend, ej: el error del regex
-    console.log("validationMessage", validationMessage);
-
-    if (!event.target?.validity?.valid) {
-      errors.push({
-        field: fieldName,
-        message: validationMessage,
-        showError: true,
-      });
-    }
-
-    setFormErrors(errors);
-    setBackendErrors(errors);
-  };
-
   return (
     <MainContainer
       onCancel={handleCancel}
@@ -106,58 +90,7 @@ export const AddApplicationArea = () => {
       hasFooterButons
       cardTitle="Alta de Campo de Aplicación"
     >
-      <div className="container">
-        {/* ------------- Nombre ------------- */}
-        <div className="row mb-3">
-          <div className="col-lg-6 col-sm-12">
-            <TextField
-              name="appAreaName"
-              required
-              fullWidth
-              inputRef={appAreaNameRef}
-              label="Nombre"
-              variant="outlined"
-              onChange={handleInputChange}
-              error={
-                formErrors.find((error) => error.field === "appAreaName")
-                  ?.showError ||
-                backendErrors.find((error) => error.field === "appAreaName")
-                  ?.showError
-              }
-              helperText={
-                formErrors.find((error) => error.field === "appAreaName")
-                  ?.message ||
-                backendErrors.find((error) => error.field === "appAreaName")
-                  ?.message
-              }
-            />
-          </div>
-          {/* ------------- Consideraciones ------------- */}
-          <div className="col-lg-6 col-sm-12">
-            <TextField
-              name="appAreaConsiderations"
-              fullWidth
-              multiline
-              inputRef={appAreaConsiderationsRef}
-              label="Consideraciones"
-              variant="outlined"
-              onChange={handleInputChange}
-              error={
-                formErrors.find((error) => error.field === "appAreaConsiderations")
-                  ?.showError ||
-                backendErrors.find((error) => error.field === "appAreaConsiderations")
-                  ?.showError
-              }
-              helperText={
-                formErrors.find((error) => error.field === "appAreaConsiderations")
-                  ?.message ||
-                backendErrors.find((error) => error.field === "appAreaConsiderations")
-                  ?.message
-              }
-            />
-          </div>
-        </div>
-      </div>
+      <ApplicationAreaForm></ApplicationAreaForm>
     </MainContainer>
   );
 };
